Agregado opcion para manejar el inventario con decimales , adicional agergada modo de facturar con notas de venta en lote , corregido problema en la pantalla de ingreso de inventario que no tiomaba en cuenta el costo y agregado hasta pv6 para las facturas

+++ workspace/inventario/src/main/java/ec/com/codesoft/codefaclite/inventario/model/StockReporteModel.java
@@ -211,7 +211,7 @@ public class StockReporteModel extends StockMinimoPanel{
                      for (Object[] objeto : listaStock) 
                      {
                         Producto producto = (Producto) objeto[0];
-                        Integer cantidad = (Integer) objeto[1];
+                        BigDecimal cantidad = (BigDecimal) objeto[1];
                         BigDecimal costoPromedio=(BigDecimal)objeto[2];
                         //Kardex kardexTemp = (Kardex) objeto[2];
